Fix Coding Style
