+++ src/Controller/AboutBaseController.php
@@ -2,16 +2,7 @@
 
 namespace App\Controller;
 
-use App\Form\FeedbackFormType;
-use App\Model\AboutModel;
-use Carbon\Carbon;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\FormError;
-use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 class AboutBaseController extends AbstractController
 {
