+++ src/Controller/AboutController.php
@@ -2,15 +2,9 @@
 
 namespace App\Controller;
 
-use App\Form\FeedbackFormType;
-use App\Model\AboutModel;
-use Carbon\Carbon;
-use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 class AboutController extends AboutBaseController
 {
