+++ src/Doctrine/SetType.php
@@ -12,7 +12,6 @@ namespace App\Doctrine;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\Type;
 use InvalidArgumentException;
-use function count;
 
 /**
  * @SuppressWarnings(PHPMD.UnusedFormalParameter)
@@ -45,9 +44,9 @@ abstract class SetType extends Type
         if (null !== $value) {
             // Split given value
             $values = explode(',', $value);
-            $valueCount = count($values);
+            $valueCount = \count($values);
 
-            if (count(array_intersect($values, $this->values)) !== $valueCount) {
+            if (\count(array_intersect($values, $this->values)) !== $valueCount) {
                 throw new InvalidArgumentException("Invalid '" . $this->name . "' value: " . $value . '.');
             }
         } else {
