+++ src/Form/GalleryUploadForm.php
@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Form;
-
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
@@ -37,9 +35,6 @@ class GalleryUploadForm extends AbstractType
             ->getForm();
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
