+++ src/Form/PasswordFormType.php
@@ -1,8 +1,6 @@
 <?php
 
-
 namespace App\Form;
-
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
