+++ src/Model/MemberDataExtractor/DoctrineExtractor.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace App\Model\MemberDataExtractor;
+
+use App\Entity\Member;
+use Doctrine\Persistence\ManagerRegistry;
+use Symfony\WebpackEncoreBundle\Asset\EntrypointLookup;
+use Twig\Environment;
+
+final class DoctrineExtractor extends AbstractExtractor implements ExtractorInterface
+{
+    private $className;
+    private $memberRelationName;
+    private $alias;
+
+    public function __construct(EntrypointLookup $entrypointLookup, Environment $environment, ManagerRegistry $registry, string $className, string $memberRelationName, string $alias)
+    {
+        parent::__construct($entrypointLookup, $environment, $registry);
+        $this->className = $className;
+        $this->memberRelationName = $memberRelationName;
+        $this->alias = $alias;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function extract(Member $member, string $tempDir): string
+    {
+        $donationRepository = $this->getRepository($this->className);
+        $donations = $donationRepository->findBy([$this->memberRelationName => $member]);
+
+        return $this->writePersonalDataFile([$this->alias => $donations], $this->alias);
+    }
+}
