+++ src/Model/MessageModel.php
@@ -538,7 +538,6 @@ class MessageModel
             'message' => $message,
             'subject' => $subjectText,
             'body' => $body,
-
         ]);
 
         return $message;
@@ -553,6 +552,36 @@ class MessageModel
         $host = $first->getReceiver();
 
         return [$thread, $first, $last, $guest, $host];
+    }
+
+    /**
+     * The requestChanged parameter triggers a PHPMD warning which is out of place in this case.
+     *
+     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
+     *
+     * @param mixed $subject
+     * @param mixed $template
+     * @param mixed $requestChanged
+     */
+    public function sendRequestNotification(
+        Member $sender,
+        Member $receiver,
+        Member $host,
+        Message $request,
+        $subject,
+        $template,
+        $requestChanged
+    ) {
+        // Send mail notification
+        $this->mailer->sendMessageNotificationEmail($sender, $receiver, $template, [
+            'host' => $host,
+            'subject' => $subject,
+            'message' => $request,
+            'request' => $request->getRequest(),
+            'changed' => $requestChanged,
+        ]);
+
+        return true;
     }
 
     private function hasLimitExceeded(Member $member, string $sql, int $perHour, int $perDay): bool
@@ -592,30 +621,4 @@ class MessageModel
 
         return false;
     }
-
-    /**
-     * The requestChanged parameter triggers a PHPMD warning which is out of place in this case
-     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
-     */
-    public function sendRequestNotification(
-        Member $sender,
-        Member $receiver,
-        Member $host,
-        Message $request,
-        $subject,
-        $template,
-        $requestChanged
-    ) {
-        // Send mail notification
-        $this->mailer->sendMessageNotificationEmail($sender, $receiver, $template, [
-            'host' => $host,
-            'subject' => $subject,
-            'message' => $request,
-            'request' => $request->getRequest(),
-            'changed' => $requestChanged,
-        ]);
-
-        return true;
-    }
-
 }
